# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "notesnet",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/internal/accounts/create": {
      "post": {
        "tags": [
          "accounts::create"
        ],
        "operationId": "create_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account created"
          },
          "409": {
            "description": "Username already in use"
          },
          "422": {
            "description": "Invalid username or password"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/internal/sessions/create": {
      "post": {
        "tags": [
          "sessions::create"
        ],
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "v1::me::get"
        ],
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Current account info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/me/scopes": {
      "get": {
        "tags": [
          "v1::me::scopes"
        ],
        "operationId": "get_me_scopes",
        "responses": {
          "200": {
            "description": "List of accessible scopes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMeScopesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid X-NotesNet-Persona header"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Persona mismatch"
          },
          "404": {
            "description": "Persona not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/notes": {
      "get": {
        "tags": [
          "v1::notes::list"
        ],
        "operationId": "list_notes",
        "parameters": [
          {
            "name": "order_by",
            "in": "query",
            "description": "Sort order field (default: written_at)",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ListNotesOrderBy"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page (default: 20, max: 100)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor (obtained from next_cursor / prev_cursor in response)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListNotesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid cursor or query parameters"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Persona mismatch"
          },
          "404": {
            "description": "Persona not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "v1::notes::create"
        ],
        "operationId": "create_note",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Note created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid X-NotesNet-Persona header"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Persona mismatch or forbidden"
          },
          "404": {
            "description": "Scope, parent note, or persona not found"
          },
          "422": {
            "description": "Duplicate parent"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/notes/{note_id}": {
      "get": {
        "tags": [
          "v1::notes::get"
        ],
        "operationId": "get_note",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "description": "Note ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListNotesItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid note_id or X-NotesNet-Persona header"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Persona mismatch"
          },
          "404": {
            "description": "Note or persona not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateNoteParent": {
        "type": "object",
        "required": [
          "parent_note_id"
        ],
        "properties": {
          "order_child": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "parent_note_id": {
            "type": "string",
            "format": "uuid"
          },
          "should_listed_as_child": {
            "type": "boolean"
          },
          "should_listed_as_parent": {
            "type": "boolean"
          }
        }
      },
      "CreateNoteRequest": {
        "type": "object",
        "required": [
          "scope_id",
          "text_for_search",
          "content_type",
          "attributes",
          "content"
        ],
        "properties": {
          "attributes": {},
          "content": {},
          "content_type": {
            "type": "string"
          },
          "parents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateNoteParent"
            }
          },
          "scope_id": {
            "type": "string",
            "format": "uuid"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "text_for_search": {
            "type": "string"
          },
          "written_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "CreateNoteResponse": {
        "type": "object",
        "required": [
          "note_id",
          "note_revision_id"
        ],
        "properties": {
          "note_id": {
            "type": "string"
          },
          "note_revision_id": {
            "type": "string"
          }
        }
      },
      "CreateSessionRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "CreateSessionResponse": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "GetMePersona": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "GetMeResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "personas"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "personas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetMePersona"
            }
          }
        }
      },
      "GetMeScopePermissions": {
        "type": "object",
        "required": [
          "can_read_note_revisions",
          "can_modify_notes",
          "can_add_their_notes_to_child"
        ],
        "properties": {
          "can_add_their_notes_to_child": {
            "type": "boolean"
          },
          "can_modify_notes": {
            "type": "boolean"
          },
          "can_read_note_revisions": {
            "type": "boolean"
          }
        }
      },
      "GetMeScopesItem": {
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_account_id",
          "is_owner",
          "permissions",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "is_owner": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "owner_account_id": {
            "type": "string"
          },
          "permissions": {
            "$ref": "#/components/schemas/GetMeScopePermissions"
          }
        }
      },
      "GetMeScopesResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetMeScopesItem"
            }
          }
        }
      },
      "ListNotesItem": {
        "type": "object",
        "required": [
          "note_id",
          "revision_id",
          "note_author_persona_id",
          "revision_author_persona_id",
          "scope_id",
          "text_for_search",
          "content_type",
          "attributes",
          "content",
          "written_at",
          "inserted_at"
        ],
        "properties": {
          "attributes": {},
          "content": {},
          "content_type": {
            "type": "string"
          },
          "inserted_at": {
            "type": "string",
            "format": "date-time"
          },
          "note_author_persona_id": {
            "type": "string"
          },
          "note_id": {
            "type": "string"
          },
          "revision_author_persona_id": {
            "type": "string"
          },
          "revision_id": {
            "type": "string"
          },
          "scope_id": {
            "type": "string"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "text_for_search": {
            "type": "string"
          },
          "written_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ListNotesOrderBy": {
        "type": "string",
        "enum": [
          "written_at",
          "inserted_at"
        ]
      },
      "ListNotesResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListNotesItem"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "prev_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}